상품관리 마무리중

+++ src/main/java/com/bookdabang/cyh/service/ProductServiceImpl.java
@@ -22,6 +22,7 @@ import com.bookdabang.cyh.domain.ProdInfo;
 import com.bookdabang.cyh.domain.ProdQnADTO;
 import com.bookdabang.cyh.domain.SearchCriteria;
 import com.bookdabang.cyh.domain.UpdateProdDTO;
+import com.bookdabang.cyh.domain.deleteProdDTO;
 import com.bookdabang.cyh.etc.API_InputProcess;
 import com.bookdabang.ljs.persistence.LoginDAO;
 
@@ -77,6 +78,23 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	@Override
+	public int deleteSelectProd(List<deleteProdDTO> list, String upPath) throws Exception {
+
+		int result = 0;
+		for (deleteProdDTO dto : list) {
+			deleteImage(upPath, dto.getImagePath());
+
+			if (pdao.deleteProd(dto.getProdNo()) != 1) {
+				result = 0;
+				break;
+			}
+			result++;
+		}
+
+		return result;
+	}
+
+	@Override
 	public boolean validationProdNo(String isbn) throws Exception {
 
 		boolean result = false;
@@ -95,11 +113,11 @@ public class ProductServiceImpl implements ProductService {
 
 	@Override
 	public boolean insertProd(InsertProdDTO product) throws Exception {
-		boolean result =false;
-		if(pdao.insertProd(product) == 1) {
+		boolean result = false;
+		if (pdao.insertProd(product) == 1) {
 			result = true;
 		}
-		
+
 		return result;
 	}
 
@@ -111,11 +129,17 @@ public class ProductServiceImpl implements ProductService {
 	}
 
 	@Override
-	public void deleteImage(String upPath, String deletePath) throws Exception {
-		System.out.println(upPath + deletePath.replace("/", File.separator));
-		File delFile = new File(upPath + deletePath.replace("/", File.separator));
+	public void deleteImage(String upPath, String imagePath) throws Exception {
 
-		delFile.delete();
+		if (imagePath.contains("http")) {
+			return;
+		} else {
+			String deletePath = imagePath.split("uploads")[1];
+			System.out.println(deletePath);
+			System.out.println(upPath + deletePath.replace("/", File.separator));
+			File delFile = new File(upPath + deletePath.replace("/", File.separator));
+			delFile.delete();
+		}
 
 	}
